Fixed relative imports to make in situ testing easier
